Removed redundant homepage.tsx component
I've decided to not make the components in the "pages" folder barebones because I think it's useless and redundant.

Instead I am going to make the components in the "pages" folder contain
all the smaller components and operate the pages main logic.

Since the home page is very minimally functional, this will be better
reflected in the play page I'm going to start working on.